+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/CustomDataProviderTest.kt
@@ -54,13 +54,14 @@ internal class CustomDataProviderTest : BaseTest() {
 
         mockServer = MockWebServer()
 
+        val searchEngineSettings = SearchEngineSettings(
+            applicationContext = targetApplication,
+            accessToken = TEST_ACCESS_TOKEN,
+            locationEngine = FixedPointLocationEngine(TEST_USER_LOCATION),
+            singleBoxSearchBaseUrl = mockServer.url("").toString()
+        )
         MapboxSearchSdk.initializeInternal(
-            searchEngineSettings = SearchEngineSettings(
-                applicationContext = targetApplication,
-                accessToken = TEST_ACCESS_TOKEN,
-                locationEngine = FixedPointLocationEngine(TEST_USER_LOCATION),
-                singleBoxSearchBaseUrl = mockServer.url("").toString()
-            ),
+            searchEngineSettings = searchEngineSettings,
             offlineSearchEngineSettings = OfflineSearchEngineSettings(
                 applicationContext = targetApplication,
                 accessToken = TEST_ACCESS_TOKEN,
@@ -74,6 +75,7 @@ internal class CustomDataProviderTest : BaseTest() {
 
         searchEngine = MapboxSearchSdk.createSearchEngine(
             apiType = ApiType.SBS,
+            searchEngineSettings = searchEngineSettings,
             coreEngine = MapboxSearchSdk.getSharedCoreEngineByApiType(ApiType.SBS),
             analyticsService = analyticsService
         )