+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/ReverseGeocodingSearchIntegrationTest.kt
@@ -72,13 +72,14 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
 
         mockServer = MockWebServer()
 
+        val searchEngineSettings = SearchEngineSettings(
+            applicationContext = targetApplication,
+            accessToken = TEST_ACCESS_TOKEN,
+            locationEngine = FixedPointLocationEngine(TEST_USER_LOCATION),
+            singleBoxSearchBaseUrl = mockServer.url("").toString()
+        )
         MapboxSearchSdk.initializeInternal(
-            searchEngineSettings = SearchEngineSettings(
-                applicationContext = targetApplication,
-                accessToken = TEST_ACCESS_TOKEN,
-                locationEngine = FixedPointLocationEngine(TEST_USER_LOCATION),
-                singleBoxSearchBaseUrl = mockServer.url("").toString()
-            ),
+            searchEngineSettings = searchEngineSettings,
             offlineSearchEngineSettings = OfflineSearchEngineSettings(
                 applicationContext = targetApplication,
                 accessToken = TEST_ACCESS_TOKEN,
@@ -92,6 +93,7 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
 
         searchEngine = MapboxSearchSdk.createSearchEngine(
             apiType = ApiType.SBS,
+            searchEngineSettings = searchEngineSettings,
             coreEngine = MapboxSearchSdk.getSharedCoreEngineByApiType(ApiType.SBS),
             analyticsService = analyticsService
         )