+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/OfflineSearchEngine.kt
@@ -65,6 +65,12 @@ public interface OfflineSearchEngine {
     }
 
     /**
+     * Experimental API, can be changed or removed in the next SDK releases.
+     * Settings used for [OfflineSearchEngine] initialization.
+     */
+    public val settings: OfflineSearchEngineSettings
+
+    /**
      * [TileStore] object used for offline tiles management.
      */
     public val tileStore: TileStore