+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/CategorySearchTest.kt
@@ -63,6 +63,7 @@ internal class CategorySearchTest {
 
         searchEngine = SearchEngineImpl(
             apiType = ApiType.GEOCODING,
+            settings = mockk(),
             analyticsService = mockk(relaxed = true),
             coreEngine = coreEngine,
             historyService = mockk(),